API

API

Creation

PairwiseListMatrices.PairwiseListMatrix — Type.

PairwiseListMatrix{T, diagonal, VT} is a (squared) symmetric matrix that stores a list of type VT with values of type T for the pairwise comparison/evaluation of nelements. If diagonal is true the first element of the list is 1, 1 otherwise is 1, 2. If diagonal is false the diagonal values are stored in a vector on the diag field.

source

Getters

PairwiseListMatrices.hasdiagonal — Function.

Returns true if the list has diagonal values.

source

PairwiseListMatrices.getlist — Function.

Retuns the list vector.

source

PairwiseListMatrices.getdiag — Function.

Retuns the diag vector (which contains the diagonal values if diagonal is false).

source

Helpers

PairwiseListMatrices.lengthlist — Function.

Returns the length of the list field

source

Returns the list length needed for a pairwise measures or comparisons of nelements. If diagonal is true, diagonal values are included in the list.

julia> plm = PairwiseListMatrix([1, 2, 3, 4, 5, 6], false)
4×4 PairwiseListMatrices.PairwiseListMatrix{Int64,false,Array{Int64,1}}:
 0  1  2  3
 1  0  4  5
 2  4  0  6
 3  5  6  0

julia> lengthlist(4, false)
6

julia> lengthlist(plm)
6

source

PairwiseListMatrices.ij2k — Function.

Returns the k index of the list from the indixes i and j with i<j from a matrix of nelements by nelements. diagonal should be true or Val{true} if the diagonal values are on the list. You must not use it with i>j.

julia> plm = PairwiseListMatrix([10,20,30,40,50,60], true)
3×3 PairwiseListMatrices.PairwiseListMatrix{Int64,true,Array{Int64,1}}:
 10  20  30
 20  40  50
 30  50  60

julia> ij2k(1, 2, 3, true)
2

julia> getlist(plm)[2]
20

source

Base.full — Function.

Returns a full dense matrix. This converts a PairwiseListMatrix{T, D, VT} into a Matrix{T}.

julia> plm = PairwiseListMatrix([10,20,30,40,50,60], true)
3×3 PairwiseListMatrices.PairwiseListMatrix{Int64,true,Array{Int64,1}}:
 10  20  30
 20  40  50
 30  50  60

julia> full(plm)
3×3 Array{Int64,2}:
 10  20  30
 20  40  50
 30  50  60

source

PairwiseListMatrices.diagonal — Function.

Returns a vector of type VT from a PairwiseListMatrix{T, false, VT} that has the diagonal values.

julia> plm = PairwiseListMatrix([10,20,30,40,50,60], true)
3×3 PairwiseListMatrices.PairwiseListMatrix{Int64,true,Array{Int64,1}}:
 10  20  30
 20  40  50
 30  50  60

julia> diagonal(plm)
3-element Array{Int64,1}:
 10
 40
 60

source

Macros

PairwiseListMatrices.@iteratelist — Macro.

The macro @iteratelist writes a for loop over the list but avoiding getfield calls inside the loop. The first argument of the macro is the PairwiseListMatrix that is going to be iterated and the second is the body of the loop. In the body list will be the list field of the PairwiseListMatrix and k the index over that list. Other variables should be interpolated in a quote. You must not modify the value of k.

julia> PLM = PairwiseListMatrix([1,2,3], false)
3x3 PairwiseListMatrices.PairwiseListMatrix{Int64,false}:
 0  1  2
 1  0  3
 2  3  0

julia> @iteratelist PLM println(list[k])
1
2
3

source

PairwiseListMatrices.@iteratediag — Macro.

The macro @iteratediag writes a for loop over the diag field of a PairwiseListMatrix{T,false} but avoiding calls to getfield inside the loop. The first argument of the macro is the PairwiseListMatrix that is going to be iterated and the second is the body of the loop. In the body diag will be the diag field of the PairwiseListMatrix and k the index over that vector. Other variables should be interpolated in a quote. You must not modify the value of k.

julia> PLM = PairwiseListMatrix([1,2,3], false)
3x3 PairwiseListMatrices.PairwiseListMatrix{Int64,false}:
 0  1  2
 1  0  3
 2  3  0

julia> @iteratediag PLM diag[k] += 10k

julia> PLM
3x3 PairwiseListMatrices.PairwiseListMatrix{Int64,false}:
 10   1   2
  1  20   3
  2   3  30

source

PairwiseListMatrices.@iterateupper — Macro.

The macro @iterateupper iterates over the upper triangular part of the PairwiseListMatrix that is given as first argument. The second argument should be true if the diagonal need to be included in the iteration or false otherwise. The last argument is the body of the loop, where list is the list and diag fields of the PairwiseListMatrix and k is the index over that list. You can also use the respective i and j indexes for that position k in the upper triangular part of the matrix. Other variables should be interpolated in a quote. You must not modify the values of i, j or k.

julia> PLM = PairwiseListMatrix([1,2,3], true)
2x2 PairwiseListMatrices.PairwiseListMatrix{Int64,true}:
 1  2
 2  3

julia> mat = zeros(Int, 2, 2)
2x2 Array{Int64,2}:
 0  0
 0  0

julia> @iterateupper PLM true :($mat)[i,j] = list[k]

julia> mat
2x2 Array{Int64,2}:
 1  2
 0  3

source

Labels/Names

PairwiseListMatrices.getlabels — Function.

It gets the labels of a PairwiseListMatrix.

julia> plm  = PairwiseListMatrix(ones(3), false)
3×3 PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}:
 0.0  1.0  1.0
 1.0  0.0  1.0
 1.0  1.0  0.0

julia> getlabels(plm)
3-element Array{String,1}:
 "1"
 "2"
 "3"

julia> nplm  = setlabels(plm, ["a","b","c"])
3×3 Named PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}
A ╲ B │   a    b    c
──────┼──────────────
a     │ 0.0  1.0  1.0
b     │ 1.0  0.0  1.0
c     │ 1.0  1.0  0.0

julia> getlabels(nplm)
3-element Array{String,1}:
 "a"
 "b"
 "c"

source

PairwiseListMatrices.setlabels — Function.

Creates a Named PairwiseListMatrix.

julia> plm  = PairwiseListMatrix(ones(3), false)
3×3 PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}:
 0.0  1.0  1.0
 1.0  0.0  1.0
 1.0  1.0  0.0

julia> nplm  = setlabels(plm, ["a","b","c"])
3×3 Named PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}
A ╲ B │   a    b    c
──────┼──────────────
a     │ 0.0  1.0  1.0
b     │ 1.0  0.0  1.0
c     │ 1.0  1.0  0.0

source

PairwiseListMatrices.setlabels! — Function.

It changes the labels of a Named PairwiseListMatrix

source

Join

Base.join — Function.

This function join two PairwiseListMatrices by their labels, returning two PairwiseListMatrices with same size and labels. There are 4 kinds of joins:

:inner : Intersect. The output matrices only include the labels that are in both PairwiseListMatrices

:outer : Union. Include the labels of the two PairwiseListMatrices.

:left : Only use labels from the first argument.

:right : Only use labels from the second argument.

NaNs are filled in where needed to complete joins. The default value for missing values can be changed passing a tuple to missing.

julia> l = setlabels(PairwiseListMatrix([1.,2.,3.], false), ["a","b","c"]) # a b c
3×3 Named PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}
A ╲ B │   a    b    c
──────┼──────────────
a     │ 0.0  1.0  2.0
b     │ 1.0  0.0  3.0
c     │ 2.0  3.0  0.0

julia> r = setlabels(PairwiseListMatrix([1.,2.,3.], false), ["b","c","d"]) # b c d
3×3 Named PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}
A ╲ B │   b    c    d
──────┼──────────────
b     │ 0.0  1.0  2.0
c     │ 1.0  0.0  3.0
d     │ 2.0  3.0  0.0

julia> join(l, r, kind=:inner) # b c
(
2×2 Named PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}
A ╲ B │   b    c
──────┼─────────
b     │ 0.0  3.0
c     │ 3.0  0.0,

2×2 Named PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}
A ╲ B │   b    c
──────┼─────────
b     │ 0.0  1.0
c     │ 1.0  0.0)

julia> join(l, r, kind=:outer) # a b c d
(
4×4 Named PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}
A ╲ B │   a    b    c    d
──────┼───────────────────
a     │ 0.0  1.0  2.0  NaN
b     │ 1.0  0.0  3.0  NaN
c     │ 2.0  3.0  0.0  NaN
d     │ NaN  NaN  NaN  NaN,

4×4 Named PairwiseListMatrices.PairwiseListMatrix{Float64,false,Array{Float64,1}}
A ╲ B │   a    b    c    d
──────┼───────────────────
a     │ NaN  NaN  NaN  NaN
b     │ NaN  0.0  1.0  2.0
c     │ NaN  1.0  0.0  3.0
d     │ NaN  2.0  3.0  0.0)

source

Statistics

sum_nodiag
mean_nodiag
zscore
zscore!

IO

PairwiseListMatrices.to_table — Function.

Creates a Matrix{Any} is useful for writedlm and/or writecsv. Labels are stored in the columns 1 and 2, and the values in the column 3. Diagonal values are included by default.

julia> plm = PairwiseListMatrix([10,20,30], false)
3×3 PairwiseListMatrices.PairwiseListMatrix{Int64,false,Array{Int64,1}}:
  0  10  20
 10   0  30
 20  30   0

julia> to_table(plm)
6×3 Array{Any,2}:
 "1"  "1"   0
 "1"  "2"  10
 "1"  "3"  20
 "2"  "2"   0
 "2"  "3"  30
 "3"  "3"   0

julia> to_table(plm, diagonal=false)
3×3 Array{Any,2}:
 "1"  "2"  10
 "1"  "3"  20
 "2"  "3"  30

source

PairwiseListMatrices.from_table — Function.

Creation of a PairwiseListMatrix from a Matrix. By default the columns with the labels for i (slow) and j (fast) are 1 and 2. Values are taken from the column 3 by default.

data = readcsv("example.csv")

3×3 Array{Any,2}:
 "A"  "B"  10
 "A"  "C"  20
 "B"  "C"  30

from_table(data, false, nothing)

3×3 Named PairwiseListMatrices.PairwiseListMatrix{Any,false,Array{Any,1}}
A ╲ B │       A        B        C
──────┼──────────────────────────
A     │ nothing       10       20
B     │      10  nothing       30
C     │      20       30  nothing

source

Base.DataFmt.writedlm — Function.

This function takes the filename as first argument and a PairwiseListMatrix as second argument. If the diagonal keyword argument is true (default), the diagonal is included in the output. The keyword argument delim (by default is ' ') allows to modified the character used as delimiter.

julia> plm  = PairwiseListMatrix(trues(3), false)
3×3 PairwiseListMatrices.PairwiseListMatrix{Bool,false,BitArray{1}}:
 false   true   true
  true  false   true
  true   true  false

julia> writedlm("example.txt", plm, diagonal=false, delim=' ')

shell> cat example.txt
1 2 true
1 3 true
2 3 true

source

Base.DataFmt.writecsv — Function.

This function takes the filename as first argument and a PairwiseListMatrix as second argument. If the keyword argument diagonal is true (default) the diagonal is included in the output.

julia> plm  = PairwiseListMatrix(trues(3), false)
3×3 PairwiseListMatrices.PairwiseListMatrix{Bool,false,BitArray{1}}:
 false   true   true
  true  false   true
  true   true  false

julia> writecsv("example.csv", plm)

shell> cat example.csv
1,1,false
1,2,true
1,3,true
2,2,false
2,3,true
3,3,false

julia> writecsv("example.csv", plm, diagonal=false)

shell> cat example.csv
1,2,true
1,3,true
2,3,true

source